Moved auth controllers to their own folder
